captcha/Binance-captcha/api/captcha_binance/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.2
Faker==16.6.1
fastapi==0.101.0
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
kiwisolver==1.4.4
loguru==0.6.0
MarkupSafe==2.1.2
matplotlib==3.7.1
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.1
numpy==1.24.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
outcome==1.2.0
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
psutil==5.9.4
pyasn1==0.4.8
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
rubicon-objc==0.4.6
scipy==1.10.1
seaborn==0.12.2
selenium==4.7.2
selenium-stealth==1.0.6
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.7.1
ultralytics==8.0.114
undetected-chromedriver==3.4.7
urllib3==1.26.13
uvicorn==0.23.2
webdriver-manager==3.8.5
websockets==10.4
Werkzeug==2.3.4
wsproto==1.2.0
zstandard==0.20.0
